Move languages json to components folder
