+++ site/gatsby-site/src/utils/updateDiscoverIndexes.js
@@ -1,6 +1,6 @@
 const config = require('../../config');
 
-const languages = require('../i18n/languages.json');
+const languages = require('../components/i18n/languages.json');
 
 const MongoClient = require('mongodb').MongoClient;
